Crude sender working

+++ src/simplebuffer.h
@@ -8,7 +8,7 @@ struct simplebuf{
   void* buffer;
   int async_writes_submitted;
   int running;
-  unsigned long file_seqnum;
+  int file_seqnum;
   int ready_to_act;
   int bufnum;
 };
